# docs/swagger.yaml
basePath: /
definitions:
  models.Song:
    properties:
      group:
        type: string
      id:
        type: integer
      link:
        type: string
      releaseDate:
        type: string
      song:
        type: string
      text:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: Поддержка API
    url: http://www.swagger.io/support
  description: Это пример сервера для управления библиотекой песен.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: API для онлайн библиотеки песен
  version: "1.0"
paths:
  /song:
    post:
      consumes:
      - application/json
      description: 'Добавление новой песни в библиотеку. Обязательные поля: group,
        song'
      parameters:
      - description: Добавить песню
        in: body
        name: song
        required: true
        schema:
          $ref: '#/definitions/models.Song'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Song'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Добавить новую песню
      tags:
      - songs
  /song/{id}:
    delete:
      consumes:
      - application/json
      description: Удаление песни из библиотеки по ID
      parameters:
      - description: ID песни
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить песню
      tags:
      - songs
    get:
      consumes:
      - application/json
      description: Получение информации о конкретной песне по ID
      parameters:
      - description: ID песни
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Song'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить информацию о песне
      tags:
      - songs
    patch:
      consumes:
      - application/json
      description: Частичное обновление данных о существующей песне по ID
      parameters:
      - description: ID песни
        in: path
        name: id
        required: true
        type: integer
      - description: Обновить поля песни
        in: body
        name: song
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Song'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Частично обновить информацию о песне
      tags:
      - songs
    put:
      consumes:
      - application/json
      description: Обновление данных о существующей песне по ID
      parameters:
      - description: ID песни
        in: path
        name: id
        required: true
        type: integer
      - description: Обновить песню
        in: body
        name: song
        required: true
        schema:
          $ref: '#/definitions/models.Song'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Song'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить информацию о песне
      tags:
      - songs
  /songs:
    get:
      consumes:
      - application/json
      description: Получение песен с пагинацией и фильтрацией
      parameters:
      - description: Номер страницы
        in: query
        name: page
        type: integer
      - description: Количество элементов на странице
        in: query
        name: limit
        type: integer
      - description: Название группы
        in: query
        name: group
        type: string
      - description: Название песни
        in: query
        name: song
        type: string
      - description: Дата выпуска песни
        in: query
        name: releaseDate
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Song'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список песен
      tags:
      - songs
swagger: "2.0"